requirements.txt
Django>=4.2.5,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.7,<2.10
drf-spectacular>=0.26.4,<0.27
requests>=2.31.0,<2.32
